Various minor patches throughout the source code to support ARM

+++ MVEE/Inc/MVEE_syscalls_support_templates.h
@@ -16,7 +16,7 @@
   The ARG<n> macros are platform-specific and implemented in MVEE_private_arch.h
 -----------------------------------------------------------------------------*/
 template<int N> constexpr 
-    unsigned long long int& arg(int variantnum)
+    unsigned long int& arg(int variantnum)
 {
 	switch(N)
 	{
